# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A lightweight, read-only Kubernetes service that provides a RESTful API for querying pod and node information\n\nThe K8s Cluster Agent runs inside a Kubernetes cluster and provides various endpoints for:\n- Pod information (description, scheduling, resources, health)\n- Node utilization metrics\n- Cluster-wide pod issues dashboard\n- Namespace error analysis",
        "title": "K8s Cluster Agent API",
        "contact": {
            "name": "K8s Cluster Agent Team",
            "url": "https://github.com/sumandas0/k8s-cluster-agent"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/cluster/pod-issues": {
            "get": {
                "description": "Returns an aggregated view of pod issues across the cluster with pattern detection and trend analysis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "Get cluster-wide pod issues",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by namespace (default: all)",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by severity (critical, warning, info)",
                        "name": "severity",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cluster issues dashboard",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.ClusterIssues"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "408": {
                        "description": "Request timeout",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/namespace/{namespace}/error": {
            "get": {
                "description": "Returns a comprehensive error analysis report for all pods in the specified namespace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Namespace"
                ],
                "summary": "Get namespace error analysis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace name",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Namespace error analysis report",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.NamespaceErrorReport"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/nodes/{nodeName}/utilization": {
            "get": {
                "description": "Returns CPU and memory utilization metrics for the specified node (requires metrics server)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nodes"
                ],
                "summary": "Get node utilization metrics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node name",
                        "name": "nodeName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Node utilization metrics",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.NodeUtilization"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Node not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Request timeout",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Metrics server not available",
                        "schema": {
                            "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.ClusterIssues": {
            "type": "object",
            "properties": {
                "calculatedAt": {
                    "type": "string"
                },
                "criticalIssues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.ClusterPodIssue"
                    }
                },
                "healthyPods": {
                    "type": "integer"
                },
                "issueCategories": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "issueVelocity": {
                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.IssueVelocity"
                },
                "issuesByNamespace": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.NamespaceIssues"
                    }
                },
                "patterns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.IssuePattern"
                    }
                },
                "topIssues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.IssueSummary"
                    }
                },
                "totalPods": {
                    "type": "integer"
                },
                "unhealthyPods": {
                    "type": "integer"
                }
            }
        },
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.ClusterPodIssue": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "containerName": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "firstSeen": {
                    "type": "string"
                },
                "isRecurring": {
                    "type": "boolean"
                },
                "lastSeen": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "nodeName": {
                    "type": "string"
                },
                "podName": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                }
            }
        },
...
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.EventInfo": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "firstTimestamp": {
                    "type": "string"
                },
                "lastTimestamp": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.IssuePattern": {
            "type": "object",
            "properties": {
                "commonLabels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "count": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "firstSeen": {
                    "type": "string"
                },
                "lastSeen": {
                    "type": "string"
                },
                "namespaces": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.IssueSummary": {
            "type": "object",
            "properties": {
                "affectedPods": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "category": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                }
            }
        },
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.IssueVelocity": {
            "type": "object",
            "properties": {
                "newIssuesLast24h": {
                    "type": "integer"
                },
                "newIssuesLastHour": {
                    "type": "integer"
                },
                "resolvedLast24h": {
                    "type": "integer"
                },
                "resolvedLastHour": {
                    "type": "integer"
                },
                "trendDirection": {
                    "type": "string"
                },
                "velocityPerHour": {
                    "type": "number"
                }
            }
        },
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.NamespaceErrorReport": {
            "type": "object",
            "properties": {
                "analysisTime": {
                    "type": "string"
                },
                "criticalIssuesCount": {
                    "type": "integer"
                },
                "healthyPodsCount": {
                    "type": "integer"
                },
                "namespace": {
                    "type": "string"
                },
                "problematicPods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.ProblematicPod"
                    }
                },
                "problematicPodsCount": {
                    "type": "integer"
                },
                "restartThresholdUsed": {
                    "type": "integer"
                },
                "summary": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.NamespaceErrorSummary"
                    }
                },
                "totalPodsAnalyzed": {
                    "type": "integer"
                },
                "warningIssuesCount": {
                    "type": "integer"
                }
            }
        },
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.NamespaceErrorSummary": {
            "type": "object",
            "properties": {
                "affectedPods": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "count": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "issueType": {
                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.PodIssueType"
                }
            }
        },
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.NamespaceIssues": {
            "type": "object",
            "properties": {
                "criticalCount": {
                    "type": "integer"
                },
                "issuesCount": {
                    "type": "integer"
                },
                "namespace": {
                    "type": "string"
                },
                "topIssues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.ClusterPodIssue"
                    }
                },
                "totalPods": {
                    "type": "integer"
                },
                "warningCount": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.NodeUtilization": {
            "type": "object",
            "properties": {
                "cpuCapacity": {
                    "type": "string"
                },
                "cpuPercentage": {
                    "type": "number"
                },
                "cpuUsage": {
                    "type": "string"
                },
                "memoryCapacity": {
                    "type": "string"
                },
                "memoryPercentage": {
                    "type": "number"
                },
                "memoryUsage": {
                    "type": "string"
                },
                "nodeName": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.PodIssue": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.PodIssueType"
                }
            }
        },
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.PodIssueType": {
            "type": "string",
            "enum": [
                "HighRestarts",
                "Pending",
                "Failed",
                "CrashLoopBackOff",
                "ImagePullError",
                "ResourceConstraints",
                "Unschedulable"
            ],
            "x-enum-varnames": [
                "PodIssueHighRestarts",
                "PodIssuePending",
                "PodIssueFailed",
                "PodIssueCrashLoop",
                "PodIssueImagePull",
                "PodIssueResourceConstraints",
                "PodIssueUnschedulable"
            ]
        },
...
        "github_com_sumandas0_k8s-cluster-agent_internal_core_models.ProblematicPod": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "issues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.PodIssue"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "ownerKind": {
                    "type": "string"
                },
                "ownerName": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                },
                "recentEvents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_core_models.EventInfo"
                    }
                },
                "restartCount": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.ErrorDetail"
                },
                "metadata": {
                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.Metadata"
                }
            }
        },
        "github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.Metadata": {
            "type": "object",
            "properties": {
                "requestId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "metadata": {
                    "$ref": "#/definitions/github_com_sumandas0_k8s-cluster-agent_internal_transport_http_responses.Metadata"
                }
            }
        },
...
    },
...
}